% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_make_table.R
\name{add_freq_var_to_dataframe}
\alias{add_freq_var_to_dataframe}
\title{add_freq_var_to_dataframe}
\usage{
add_freq_var_to_dataframe(raw_data, data, freq_var)
}
\description{
add_freq_var_to_dataframe
}
