% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordeling.R
\name{add_gjen_var_to_dataframe}
\alias{add_gjen_var_to_dataframe}
\title{add_gjen_var_to_dataframe}
\usage{
add_gjen_var_to_dataframe(raw_data, data, gjen_var)
}
\description{
add_gjen_var_to_dataframe
}
