% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Def_henteData.R
\name{alleRegData}
\alias{alleRegData}
\title{Hente og koble sammen "alle" tabeller
Ferdigstilte data}
\usage{
alleRegData(egneVarNavn = 0)
}
\description{
Hente og koble sammen "alle" tabeller
Ferdigstilte data
}
