% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{boxplot_sam}
\alias{boxplot_sam}
\title{Lag boxplot}
\usage{
boxplot_sam(data, gg_data, input_data)
}
\arguments{
\item{data}{datasett som skal plottes}

\item{input_data}{datasett med brukerens UI-valg}

\item{ggdata}{datasett med navn til ggplot}
}
\value{
boxplot
}
\description{
Lag boxplot
}
