% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{check_small_sample}
\alias{check_small_sample}
\title{Check for small sample size}
\usage{
check_small_sample(data, var1, var2)
}
\value{
a list of data frames
}
\description{
Check for small sample size
}
