% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_datadump.R
\name{clean_datadump}
\alias{clean_datadump}
\title{Clean_datadump}
\usage{
clean_datadump(
  data,
  dato1,
  dato2,
  kjoenn,
  alder1,
  alder2,
  userRole,
  userUnitId
)
}
\arguments{
\item{data}{datasett}

\item{dato1}{brukervalg - dato min}

\item{dato2}{brukervalg - dato max}

\item{kjoenn}{brukervalg - kjønn}

\item{alder1}{brukervalg - alder min}

\item{alder2}{brukervalg - alder max}

\item{userRole}{brukerrolle}

\item{userUnitId}{brukertilhørighet}
}
\value{
datasett filtrert på brukervalg
}
\description{
Clean data for module data dump
}
