% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{comparison_plot_discrete}
\alias{comparison_plot_discrete}
\title{Plot for comparison

returns a plot}
\usage{
comparison_plot_discrete(data1, data2, labels, comp1, comp2)
}
\description{
Plot for comparison

returns a plot
}
