% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_kvalitetsindikatorer.R
\name{count_kvalind}
\alias{count_kvalind}
\title{Kvalitetsindikatortelling}
\usage{
count_kvalind(data, kjoenn, type_op, var, userRole, userUnitId, map_data)
}
\description{
Kvalitetsindikatortelling
}
