% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Def_fordelsfigurOgTab.R
\name{DefFordeling}
\alias{DefFordeling}
\title{Title}
\usage{
DefFordeling(
  RegData,
  valgtVar = "alder",
  datoFra = "2023-01-01",
  datoTil = Sys.Date(),
  aar = 0,
  minald = 0,
  maxald = 110,
  erMann = "",
  outfile = "",
  reshID = 0,
  enhetsUtvalg = 0
)
}
\arguments{
\item{RegData}{preprosessert dataramme}

\item{valgtVar}{parameter som angir hvilke(n) variabel man ønsker å
tilrettelegge for videre beregning.}

\item{datoFra}{Operasjonsdato, fra og med. Standard: '2023-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: Sys.Date()}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: 9
Alt annet enn 0 og 1 gir begge}

\item{outfile}{utfil for figur}

\item{enhetsUtvalg}{Sammenlikning eller ikke: 0-hele landet, 1-egen enhet mot resten av landet, 2-egen enhet}
}
\value{
Fordelingsfigur
}
\description{
Title
}
