% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_henteData.R
\name{defHentData}
\alias{defHentData}
\title{Hent datatabell fra Deformitets database}
\usage{
defHentData(tabellnavn = "surgeonform")
}
\arguments{
\item{tabellnavn}{Navn på tabell som skal lastes inn.
Kan ha følgende verdier:
mce
mce_patient_data
patient
patientfollowup
patientform
surgeonfollowup
surgeonform}
}
\description{
Hent datatabell fra Deformitets database
}
