% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{density_plot_comparability}
\alias{density_plot_comparability}
\title{Plot for comparison}
\usage{
density_plot_comparability(data, labels, input_data)
}
\value{
a density plot
}
\description{
Plot for comparison
}
