% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{density_sam}
\alias{density_sam}
\title{Density plot for sammenligning}
\usage{
density_sam(data, gg_data, input_data)
}
\arguments{
\item{data}{datasett som skal plottes}

\item{gg_data}{datasett med fine navn til ggplot}

\item{input_data}{datasett med brukerens UI-valg}
}
\value{
density plot (tetthetsplot)
}
\description{
Density plot for sammenligning
}
