% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_figAndelerGrVar.R
\name{figAndelerGrVar}
\alias{figAndelerGrVar}
\title{Søylediagram med andeler for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
figAndelerGrVar(
  RegData = 0,
  hentData = 0,
  preprosess = 0,
  valgtVar = "liggetidPostOp",
  minald = 0,
  maxald = 130,
  erMann = 9,
  datoFra = "2023-01-01",
  datoTil = Sys.Date(),
  Ngrense = 10,
  reshID = 0,
  outfile = "",
  ...
)
}
\arguments{
\item{RegData}{dataramme}

\item{hentData}{hente data? 0-nei, 1-ja}

\item{preprosess}{preprosessere data? 0-nei, 1-ja}

\item{valgtVar}{Variabelen det skal vises resultat for.}

\item{minald}{minimum alder}

\item{maxald}{maksimum alder}

\item{erMann}{kjønn 1=mann, 0=kvinne}

\item{datoFra}{startdato}

\item{datoTil}{sluttdato}

\item{Ngrense}{Minste antall registreringer for at ei gruppe skal bli vist}

\item{reshID}{avdelingas reshid}

\item{outfile}{Format på figurfila: pdf, png,...
''-standard, viser figur på skjerm}

\item{...}{mulige tilleggsvariabler}
}
\value{
Figur som viser andel av en variabel for gitt grupperingsvariabel.
Grupperingsvariabelen (grVar) er som regel den enkelte sykehusenhet.
}
\description{
Funksjon som genererer en figur med andeler av en variabel for en valgt
grupperingsvariabel, foreløpig bare sykehus.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable
enn sykehus
}
