% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Def_FiltrerUtvalg.R
\name{filtrer_datadump}
\alias{filtrer_datadump}
\title{Filtrer datadump}
\usage{
filtrer_datadump(data, dato1, dato2, userRole, userUnitId)
}
\arguments{
\item{data}{datasett}

\item{dato1}{brukervalg - dato min}

\item{dato2}{brukervalg - dato max}

\item{userRole}{brukerrolle}

\item{userUnitId}{brukertilhørighet}
}
\value{
datasett filtrert på brukervalg
}
\description{
Filtrer data for module data dump
Funksjonen filtrerer data
}
