% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{find_comp_variables}
\alias{find_comp_variables}
\title{Find exact variables}
\usage{
find_comp_variables(data, choice_comp)
}
\value{
two variables chosen by UI
}
\description{
Find exact variables
}
