% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{find_variables}
\alias{find_variables}
\title{Find variables over time}
\usage{
find_variables(var)
}
\value{
values that can be used for selection
}
\description{
Find variables over time
}
