% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{finn_sam_konfidensint}
\alias{finn_sam_konfidensint}
\title{Funksjon som regner gjennomsnitt og konfidensintervall for variabler}
\usage{
finn_sam_konfidensint(data)
}
\arguments{
\item{data}{datasett som har vært gjennom vask_sam_tabell()}
}
\value{
datasett med konfidensintervall for variabelen valgt i ui-delen
}
\description{
Funksjon som regner gjennomsnitt og konfidensintervall for variabler
}
