% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{finn_sam_variabler}
\alias{finn_sam_variabler}
\title{Finn to variabler}
\usage{
finn_sam_variabler(data, valg_sam)
}
\arguments{
\item{data}{datasett som skal brukes}

\item{valg_sam}{brukerens UI-valg av variabler som skal sammenlignes}
}
\value{
datasett med to variabler som skal sammenlignes
}
\description{
Finn to variabler
}
