% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeHist_ggplot.R
\name{gg_makeHist}
\alias{gg_makeHist}
\title{ggplot2 function for use in Rapporteket
the function creates bar plots}
\usage{
gg_makeHist(data, x_var, q, w, l)
}
\arguments{
\item{data}{dataframe from which output is to be made}

\item{x_var}{defining which variable in the data frame to be used to plot (on the x-axis)}

\item{makeTable}{Logical that if TRUE function will return a data frame
containing the bin borders and count within each bin}
}
\value{
a graphical object or data frame
}
\description{
ggplot2 function for use in Rapporteket
the function creates bar plots
}
\examples{
gg_makeHist(data = regdata, x_var = BMI_CATEGORY) # provide without quotationmarks
}
