% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{ggdata_sam_plot}
\alias{ggdata_sam_plot}
\title{Data for gg-plotting}
\usage{
ggdata_sam_plot(var)
}
\arguments{
\item{var}{Variabel som skal få finere navn}
}
\value{
datasett med fine navn til ggplot
}
\description{
Data for gg-plotting
}
