% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordeling_FASESut.R
\name{gjen_var_til_data}
\alias{gjen_var_til_data}
\title{gjen_var_til_data}
\usage{
gjen_var_til_data(raw_data, data, gjen_var)
}
\arguments{
\item{raw_data}{rå-RegData}

\item{data}{RegData som har vært gjennom prepros}

\item{gjen_var}{valgt variabel}
}
\description{
Denne funksjonen tar ei kolonne fra ei dataramme og legger den til
ei annen dataramme. Dette gjøres pr. forløps-id. Kolonnen som legges til
får navnet "gjen_var"
}
\examples{
\donttest{
try(gjen_var_til_data(raw_regdata, RegData, "Diff_prosent_kurve"))
}

}
