% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpeFunksjoner.R
\name{kjor_app_deform}
\alias{kjor_app_deform}
\title{Run the Deformitet Shiny Application}
\usage{
kjor_app_deform()
}
\value{
An object representeing the Deformitet app
}
\description{
Run the Deformitet Shiny Application
}
