% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_deform.R
\name{kjorDeformApp}
\alias{kjorDeformApp}
\title{Run the Deformitet Shiny Application}
\usage{
kjorDeformApp(browser = FALSE, logAsJson = FALSE)
}
\arguments{
\item{browser}{Open app in browser window}

\item{logAsJson}{Log in json-format}
}
\value{
An object representeing the Deformitet app
}
\description{
Run the Deformitet Shiny Application
}
