% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_count_kvalitetsindikator.R
\name{kval_count}
\alias{kval_count}
\title{Calculate counts for kvalitetsiindikator}
\usage{
kval_count(data, var, kjønn, type_op)
}
\description{
Calculate counts for kvalitetsiindikator
}
