% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordeling_FASESut.R
\name{lag_gjen_tabell}
\alias{lag_gjen_tabell}
\title{lag_gjen_tabell}
\usage{
lag_gjen_tabell(data)
}
\arguments{
\item{data}{data som har vært gjennom prepVar()}
}
\description{
lag_gjen_tabell
}
\examples{
\donttest{
try(lag_gjen_tabell(data))
}

}
