% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Def_fordelsfigurOgTab.R
\name{lagFordelingstabell}
\alias{lagFordelingstabell}
\title{Beregne antall og andeler til fordelingsfigur/tabell}
\usage{
lagFordelingstabell(RegData, flerevar, variabler = NULL, ind)
}
\arguments{
\item{RegData}{Dataramme med preprossesserte datta}

\item{flerevar}{fordeling av en eller flere variabler (0 el 1)}

\item{variabler}{hvilke variabler som er med}

\item{ind}{indeks for hvilke registreringer som er med i $Hovedgruppa og
hvilke som er med i resten $Rest}
}
\value{
Beregnede andeler, samt info om grupper
}
\description{
Beregne antall og andeler til fordelingsfigur/tabell
}
