% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordeling_FASESut.R
\name{lagFordelingstabellGml}
\alias{lagFordelingstabellGml}
\title{lagFordelingstabell}
\usage{
lagFordelingstabellGml(data, var_reshid = 0, visning = "Hele landet")
}
\arguments{
\item{data}{data som har vært gjennom prepVar()}

\item{var_reshid}{reshID (valg kun tilgj. som SC-bruker)}

\item{visning}{valg av visning}
}
\description{
lagFordelingstabell
}
\examples{
\donttest{
try(lagFordelingstabell(data, 103240, "egen enhet"))
}

}
