% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{make_comparability_table}
\alias{make_comparability_table}
\title{Make table}
\usage{
make_comparability_table(data, var)
}
\value{
a dataframe that is in the long format
}
\description{
Make table
}
