% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{make_labels}
\alias{make_labels}
\title{Make labels

returns a dataframe with conditional labels}
\usage{
make_labels(data1, data2, comp1, comp2)
}
\description{
Make labels

returns a dataframe with conditional labels
}
