% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTable.R
\name{make_table}
\alias{make_table}
\title{Table function for calculating proportions by hospital
and by choice of two variable - intention is to use "GENDER" given the set up
of the shiny application + one choice of variable (e.g., BMI_CATEGORY)}
\usage{
make_table(data, my_var, my_var2)
}
\arguments{
\item{data}{: data frame}

\item{my_var}{: the variable chosen from the small drop down menu -
gender: both, men, women}

\item{my_var2}{: the variable chosen from the long drop down menu with all variables
that you can want to see proportion analyses for}
}
\value{
a data frame
}
\description{
Table function for calculating proportions by hospital
and by choice of two variable - intention is to use "GENDER" given the set up
of the shiny application + one choice of variable (e.g., BMI_CATEGORY)
}
\examples{
takes a data frame and (i) preforms selections based on input; (ii)
calculates proportions and percentages; (iii) returns a data frame

}
