% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_make_ggplot.R
\name{makePlot_gg}
\alias{makePlot_gg}
\title{Make gg_plot for fordelingsfirgurer og -tabeller}
\usage{
makePlot_gg(data, gg_data, data_var)
}
\description{
Make gg_plot for fordelingsfirgurer og -tabeller
}
