% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Def_hjelpefunksj.R
\name{mapping_navn}
\alias{mapping_navn}
\title{mapping_navn}
\usage{
mapping_navn(data, x_var)
}
\arguments{
\item{data}{rådata (data som ikke har vært gjennom prepros)}

\item{x_var}{valgt variabel i UI-delen}
}
\description{
Denne funksjonen lager mapping mellom variabler som har vært gjennom prePros()
og variabler som kommer fra rådataen
}
\examples{
\donttest{
try(mapping_navn(raw_regdata, "BMI_kategori"))
}

}
