% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_les_og_flate_ut.R
\name{mergeRegData}
\alias{mergeRegData}
\title{Les inn data og lag utflatet dataramme}
\usage{
mergeRegData(
  mce,
  centre,
  mce_patient_data,
  patient,
  patient_followup,
  patient_form,
  surgeon_followup,
  surgeon_form
)
}
\description{
Les inn data og lag utflatet dataramme
}
