% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_fordeling.R
\name{module_fordeling_server}
\alias{module_fordeling_server}
\title{Server fordeling}
\usage{
module_fordeling_server(id, userRole, userUnitId, regdata, raw_data, map_data)
}
\description{
Server fordeling
}
