% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_gjennomsnitt.R
\name{module_gjennomsnitt_server}
\alias{module_gjennomsnitt_server}
\title{Server gjennomsnitt}
\usage{
module_gjennomsnitt_server(id, userRole, userUnitId, data, map_data)
}
\description{
Server gjennomsnitt
}
