% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_registreringer.R
\name{module_registreringer_server}
\alias{module_registreringer_server}
\title{Server sammenligningsmodul}
\usage{
module_registreringer_server(id, userRole, userUnitId, data)
}
\description{
Server sammenligningsmodul
}
