% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_figAndelerGrVar.R
\name{NakkeFigAndelerGrVar}
\alias{NakkeFigAndelerGrVar}
\title{Søylediagram med andeler for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NakkeFigAndelerGrVar(
  RegData = 0,
  hentData = 0,
  preprosess = 0,
  valgtVar = "Alder",
  minald = 0,
  maxald = 130,
  datoFra = "2023-01-01",
  datoTil = Sys.Date(),
  erMann = "",
  Ngrense = 10,
  reshID = 0,
  outfile = "",
  ...
)
}
\arguments{
\item{RegData}{dataramme}

\item{hentData}{hente data? 0-nei, 1-ja}

\item{preprosess}{preprosessere data? 0-nei, 1-ja}

\item{valgtVar}{Variabelen det skal vises resultat for.}

\item{minald}{minimum alder}

\item{maxald}{maksimum alder}

\item{Ngrense}{Minste antall registreringer for at ei gruppe skal bli vist}

\item{...}{mulige tilleggsvariabler}
}
\value{
Figur som viser andel av en variabel for gitt grupperingsvariabel.
Grupperingsvariabelen (grVar) er som regel den enkelte sykehusenhet.
}
\description{
Funksjon som genererer en figur med andeler av en variabel for en valgt
grupperingsvariabel, foreløpig bare sykehus.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable
enn sykehus
}
