% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{nye_navn}
\alias{nye_navn}
\title{Lage nye navn på kolonne}
\usage{
nye_navn(data)
}
\arguments{
\item{data}{datasett som skal endres}
}
\value{
datasett med nye kolonnenavn
}
\description{
Lage nye navn på kolonne
}
