% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gjennomsnitt.R
\name{over_tid_plot}
\alias{over_tid_plot}
\title{Plot - gjennomsnitt over tid}
\usage{
over_tid_plot(data, visning, gg_data, map_var, tidsenhet, data_var)
}
\arguments{
\item{data}{data som har vært gjennom tabell_gjen_tid()}

\item{visning}{valgt visning}

\item{gg_data}{data fra prepVar()[2]}

\item{map_var}{data som mapper mellom variabler (faktornivå og "kontinuerlig variabel")}

\item{tidsenhet}{valgt tidsenhet}

\item{data_var}{datasett med brukerens UI-valg}
}
\description{
Plot - gjennomsnitt over tid
}
\examples{
\donttest{
over_tid_plot(data, "egen enhet", gg_data, map_var, "kvartal")
}

}
