% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Def_preprosessData.R
\name{preprosData}
\alias{preprosData}
\title{Preprosesser data fra NKR Deformitet}
\usage{
preprosData(RegData = RegData, egneVarNavn = 0)
}
\arguments{
\item{RegData}{dataramme med registerets data}
}
\value{
RegData En dataramme med det preprosesserte datasettet
}
\description{
Denne funksjonen definerer og formaterer variabler
}
