% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Def_preprosessData.R
\name{prepVar}
\alias{prepVar}
\title{prepvar-funksjon. Lar denne stå som den er, men lager filtrerings og
tilretteleggingsfunksjoner ut fra vanlig struktur. SKAL FASES UT}
\usage{
prepVar(
  data,
  var,
  var_kjonn,
  time1,
  time2,
  alder1,
  alder2,
  type_op = "Begge",
  visning = "uten_tid"
)
}
\value{
Funksjonen ser ut tiil både å filtrere og preprosessere
}
\description{
prepvar-funksjon. Lar denne stå som den er, men lager filtrerings og
tilretteleggingsfunksjoner ut fra vanlig struktur. SKAL FASES UT
}
