% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_deform.R
\name{server_deform}
\alias{server_deform}
\title{Server logic for the deformitet app}
\usage{
server_deform(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
A shiny app server object
}
\description{
Server logic for the deformitet app
}
