% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gjennomsnitt.R
\name{sjekk_antall}
\alias{sjekk_antall}
\title{Sjekke antall
Funksjon for å sjekke størrelse på datagrunnlaget. Funksjonen gjør at man
dropper å vise variabelen om det ikke finnes noen registreringer på gitt variabel
i hoveddataen og om det er na i datasettet som er regnet om med tabell_gjen_tid()}
\usage{
sjekk_antall(data, data1, date1, date2, tidsenhet)
}
\arguments{
\item{data}{datasett som ikke har blitt omgjort}

\item{data1}{datasett som har vært gjennom tabell_gjen_tid()}

\item{date1}{brukerens valg for første dato}

\item{date2}{brukerens valg for siste dato}

\item{tidsenhet}{brukerens valg av tidsenhet (kvartal vs. år)}
}
\description{
Sjekke antall
Funksjon for å sjekke størrelse på datagrunnlaget. Funksjonen gjør at man
dropper å vise variabelen om det ikke finnes noen registreringer på gitt variabel
i hoveddataen og om det er na i datasettet som er regnet om med tabell_gjen_tid()
}
\examples{
\donttest{
try(sjekk_antall(RegData, tabell_gjen_tid, "2024-01-01", "2025-01-01", "aar"))
}
}
