% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gjennomsnitt.R
\name{tabell_gjen_tid}
\alias{tabell_gjen_tid}
\title{Tabell - gjennomsnitt over tid}
\usage{
tabell_gjen_tid(
  data,
  var,
  map_data,
  tidsenhet = "kvartal",
  visning = "hele landet",
  userUnitId
)
}
\arguments{
\item{data}{data som har vært gjennom mapping_navn() og prepVar()}

\item{var}{brukervalg av variabel}

\item{map_data}{mapping-data mellom reshId og sykehusnavn}

\item{tidsenhet}{brukervalg av tid}

\item{visning}{brukervalg av visning}

\item{userUnidIt}{brukerens enhetstilhørighet}
}
\description{
Tabell - gjennomsnitt over tid
}
\examples{
tabell <- tabell_gjen_tid(prep_data, var, map_db_resh, "kvartal", "hver enhet", 111961)

}
