% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{tabell_sam_discrete}
\alias{tabell_sam_discrete}
\title{Sammenligningstabell - grupper av faktorer}
\usage{
tabell_sam_discrete(data, comp1, comp2)
}
\description{
Sammenligningstabell - grupper av faktorer
}
