% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_tid.R
\name{table_freq_time}
\alias{table_freq_time}
\title{Tabell - gjennomsnitt over tid}
\usage{
table_freq_time(
  data,
  var,
  map_data,
  tidsenhet = "kvartal",
  visning = "hele landet",
  userUnitId
)
}
\description{
Tabell - gjennomsnitt over tid
}
