% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Def_Tabeller.R
\name{tbl_skjema_reg}
\alias{tbl_skjema_reg}
\title{tabell registreringer pr skjema}
\usage{
tbl_skjema_reg(date1, date2, data)
}
\arguments{
\item{date1}{dato-valg (min) gjort av brukeren}

\item{date2}{dato-valg (max) gjort av brukeren}

\item{data}{data -> data som har vært gjennom prePros()}
}
\description{
tabell registreringer pr skjema
}
\examples{
\donttest{
try(tbl_skjema_reg("01-09-2024", "01-01-2025", RegData))
}

}
