% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Def_FiltrerUtvalg.R
\name{utvalgEnh}
\alias{utvalgEnh}
\title{Funksjon som gjør utvalg i datagrunnlaget til registreringene}
\usage{
utvalgEnh(
  RegData,
  datoFra = "2023-01-01",
  datoTil = Sys.Date(),
  minald = 0,
  maxald = 110,
  erMann = 9,
  aar = 0,
  enhetsUtvalg = 0,
  reshID = 0,
  fargepalett = "BlaaOff"
)
}
\arguments{
\item{datoFra}{Operasjonsdato, fra og med. Standard: '2023-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: Sys.Date()}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: 9
Alt annet enn 0 og 1 gir begge}

\item{enhetsUtvalg}{Sammenlikning eller ikke: 0-hele landet, 1-egen enhet mot resten av landet, 2-egen enhet}

\item{fargepalett}{- Velge fargepalett, standard:BlaaOff ("offentliggjøringsfargene")}
}
\description{
Funksjon som gjør utvalg i datagrunnlaget til registreringene
}
