% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{vask_sam_tabell}
\alias{vask_sam_tabell}
\title{Vaske datasettet}
\usage{
vask_sam_tabell(data, var)
}
\arguments{
\item{data}{datasettet som skal vaskes}

\item{var}{variabel brukeren har valgt i UI-delen}
}
\value{
et vasket datasett
}
\description{
Vaske datasettet
}
