% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gjennomsnitt.R
\name{y_limits_gjen}
\alias{y_limits_gjen}
\title{Y-axis limits
Funksjon som setter fornuftige øvre og nedre grenser på yaksen.}
\usage{
y_limits_gjen(var)
}
\arguments{
\item{var}{variabel som velges av bruker i UI-delen}
}
\description{
Y-axis limits
Funksjon som setter fornuftige øvre og nedre grenser på yaksen.
}
\examples{
\donttest{
ylimits_gjen(SRS22_MAIN_SCORE)
}

}
