% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{check_small_sample}
\alias{check_small_sample}
\title{Check for small sample size

Returns a list of data frames}
\usage{
check_small_sample(var1, var2)
}
\description{
Check for small sample size

Returns a list of data frames
}
