% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_kvalitetsindikatorer.R
\name{count_kvalind}
\alias{count_kvalind}
\title{Kvalitetsindikatortelling}
\usage{
count_kvalind(data, kjoenn, var, userRole, userUnitId, map_data)
}
\arguments{
\item{data}{data som har vært gjennom prepVar()}

\item{kjoenn}{brukerens valg på kjønn}

\item{var}{brukerens valg av variabel}

\item{userRole}{brukerens brukerrolle}

\item{userUnitId}{brukuers reshId}

\item{map_data}{datasett som mapper reshId og sykehusnavn}
}
\description{
Funksjon som teller tilfeller av visse vilkår til kvalitetsindikatorer
}
\examples{
\donttest{
try(count_kvalind(regdata, "begge", "PRE_MAIN_CURVE", "SC", 111961, map_db_resh))
}

}
