% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_les_og_flate_ut.R
\name{deformitetHentTabell}
\alias{deformitetHentTabell}
\title{Hent tabell fra Deformitets database}
\usage{
deformitetHentTabell(tabellnavn = "surgeonform")
}
\arguments{
\item{tabellnavn}{Navn på tabell som skal lastes inn.
Kan ha følgende verdier:
mce
mce_patient_data
patient
patientfollowup
patientform
surgeonfollowup
surgeonform}
}
\description{
Hent tabell fra Deformitets database
}
