% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_kvalitetsindikatorer.R
\name{explanation_kvalind}
\alias{explanation_kvalind}
\title{Function for getting explanations for kvalitetsindikatorer}
\usage{
explanation_kvalind(kjønn_choice, var)
}
\arguments{
\item{kjønn_choice}{brukervalg på visning kjønn}

\item{var}{brukervalg på variabel}
}
\description{
Function for getting explanations for kvalitetsindikatorer
}
\examples{
\donttest{
try(explanation_kvalind("begge", "SRS22_spm22_3mnd"))
}

}
