% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_explanation_kvalind.R,
%   R/func_kvalitetsindikatorer.R
\name{explanation_kvalind}
\alias{explanation_kvalind}
\title{Function for getting explanations for kvalitetsindikatorer}
\usage{
explanation_kvalind(kjønn_choice, var)

explanation_kvalind(kjønn_choice, var)
}
\description{
Function for getting explanations for kvalitetsindikatorer

Function for getting explanations for kvalitetsindikatorer
}
