% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{finn_variabler}
\alias{finn_variabler}
\title{Finn variabler over tid}
\usage{
finn_variabler(var)
}
\arguments{
\item{var}{input fra UI - valg av variabel}
}
\value{
en vektor med navn på kolonner som hører til valgt variabel
}
\description{
Finn variabler over tid
}
