% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordeling.R
\name{gjen_var_til_data}
\alias{gjen_var_til_data}
\title{gjen_var_til_data}
\usage{
gjen_var_til_data(raw_data, data, gjen_var)
}
\arguments{
\item{raw_data}{rå-regdata}

\item{data}{regdata som har vært gjennom prepros}

\item{gjen_var}{valgt variabel}
}
\description{
Denne funksjonen tar ei kolonne fra ei dataramme og legger den til
ei annen dataramme. Dette gjøres pr. forløps-id. Kolonnen som legges til
får navnet "gjen_var"
}
\examples{
check <- gjen_var_til_data(raw_regdata, regdata, "Diff_prosent_kurve")

}
