% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_komplikasjonstype.R
\name{kompl_data}
\alias{kompl_data}
\title{Komplikasjonstyper}
\usage{
kompl_data(
  regdata,
  var,
  var_kjønn,
  time1,
  time2,
  alder1,
  alder2,
  type_op,
  map_data
)
}
\description{
Komplikasjonstyper
}
