% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_kvalitetsindikatorer.R
\name{kval_plot}
\alias{kval_plot}
\title{Funksjon som lager kvalitetsindikatorplot}
\usage{
kval_plot(data, gg_data, data_var, choice_kjonn)
}
\arguments{
\item{data}{data som har vært gjennom count_kvalind()}

\item{gg_data}{data som kommer fra perpVar()[2]}

\item{choice_kjonn}{brukervalg på visning på kjønn}
}
\description{
Funksjon som lager kvalitetsindikatorplot
}
\examples{
\donttest{
try(kval_plot(data, gg_data, data_var, "nei"))
}

}
