% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordelsfigur_ggplot.R
\name{lag_ggplot_fordeling}
\alias{lag_ggplot_fordeling}
\title{Make gg_plot for fordelingsfirgurer og -tabeller}
\usage{
lag_ggplot_fordeling(data, gg_data, data_var, visning)
}
\arguments{
\item{data}{data som kommer ut fra lagTabell()}

\item{gg_data}{gg-data som kommer fra prepVar()}

\item{data_var}{data som lagrer brukerens valg}

\item{visning}{valg av visning}
}
\description{
Make gg_plot for fordelingsfirgurer og -tabeller
}
\examples{
lag_ggplot_fordeling(data, gg_data, data_var, "egen enhet")
}
