% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sammenligningsmodul.R
\name{lag_sam_tabell}
\alias{lag_sam_tabell}
\title{Lag tabell til sammenligning
Denne funksjonen bruker funksjonen over til å lage et datasett/tabell med
alle kolonnene av interesse i et langt format. Det er kun variabelene som er
valgt, samt Sykehus som blir med i funksjonen.}
\usage{
lag_sam_tabell(data, var)
}
\arguments{
\item{data}{datasett}

\item{var}{variabelen som skal brukes i finn_variabler()}
}
\value{
datasett/tabell i langt format
}
\description{
Lag tabell til sammenligning
Denne funksjonen bruker funksjonen over til å lage et datasett/tabell med
alle kolonnene av interesse i et langt format. Det er kun variabelene som er
valgt, samt Sykehus som blir med i funksjonen.
}
\examples{
\donttest{
try(lag_sam_tabell(RegData, "Funksjon"))
}

}
