% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_henteData.R
\name{mappingEgneNavnDum}
\alias{mappingEgneNavnDum}
\title{Endre variabelnavn/kolonnenavn til selvvalgte navn}
\usage{
mappingEgneNavnDum(tabell, tabType)
}
\arguments{
\item{tabell}{datatabellnavn i databasen}

\item{tabType}{REGISTRATION_TYPE}
}
\value{
tabell med selvvalgte variabelnavn spesifisert i friendlyvar
}
\description{
Endre variabelnavn/kolonnenavn til selvvalgte navn
}
