% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prep_data.R
\name{prepVar}
\alias{prepVar}
\title{prepvar-funksjon}
\usage{
prepVar(
  data,
  var,
  var_kjonn,
  time1,
  time2,
  alder1,
  alder2,
  type_op,
  visning = "uten_tid"
)
}
\description{
prepvar-funksjon
}
