% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_filtrerUtvalg.R
\name{utvalg_basic}
\alias{utvalg_basic}
\title{Utvalgsfunksjon}
\usage{
utvalg_basic(
  data,
  user_unit,
  gender,
  type_op,
  tid1,
  tid2,
  alder1,
  alder2,
  bruk_av_funk
)
}
\description{
Utvalgsfunksjon
}
