% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Def_TilretteleggeVar.R
\name{varTilrettelegg}
\alias{varTilrettelegg}
\title{Funksjon for å tilrettelegge variabler for beregning.}
\usage{
varTilrettelegg(RegData, valgtVar, figurtype = "andeler")
}
\arguments{
\item{valgtVar}{parameter som angir hvilke(n) variabel man ønsker å
tilrettelegge for videre beregning.}

\item{figurtype}{Hvilken figurtype det skal tilrettelegges variabler for:
'andeler', 'andelGrVar', 'andelTid', 'gjsnGrVar', 'gjsnTid'}
}
\value{
Definisjon av valgt variabel.
}
\description{
Funksjonen gjør utvalg og tilrettelegger variabler (gitt ved valgtVar) til
videre bruk. Videre bruk kan eksempelvis være beregning av aggregerte
verdier.
Funksjonen gjør også filtreringer som å fjerne ugyldige verdier for den
valgte variabelen, samt ta høyde for avhengigheter med
andre variabler. Det er også her man angir aksetekster og titler for den
valgte variabelen.
NB: Mange variabler er tilrettelagt i preprosseseringa, arv fra Ingrid...
}
